feat: rewrite CI, packaging and examples for the native build
CI drops everything Rust and gains the jobs the new implementation needs:
clang-format and clang-tidy, a build-and-test matrix across Linux, macOS and
Windows, ASan with UBSan, an 80% line-coverage gate, a 60-second libFuzzer
smoke over every parser, and an install-and-consume step so find_package is
proven rather than assumed.

The zero-dependency job is the one that matters. It deliberately installs no
libcurl: with TAMGA_HTTP=none the library must build, pass the whole suite,
and link against libc alone, asserted by inspecting ldd output. If that job
ever needs a package installed, the claim on the tin has stopped being true.

Static analysis found real defects, all fixed rather than suppressed:

- two null-pointer dereferences on paths where a successful call's response
  was assumed non-null before being dereferenced (activate_machine reading
  back the created machine's id, and has_entitlement reading the listing);
- a null pointer reaching memcpy in the buffer's growth path;
- two dead stores, one of which sat in the Montgomery inner loop where it
  read as meaningful;
- a dead initialiser in the merge sort that was assigned before a clamp it
  needed to follow;
- seven implicit widening multiplications used as pointer offsets, all in
  bounded loops and none reachable, now structurally safe via size_t indices;
- a duplicated branch in URL construction and a missing-parentheses
  precedence warning in the civil-date conversion.

The two remaining findings are name-similarity false positives; each is
annotated at its own site with why the argument order is correct, rather than
disabling a check that catches real swaps elsewhere.

Versioning: release-please now drives the four places the version appears
(CMakeLists.txt, include/tamga.h, and both vcpkg.json files) through
extra-files. That drift was already real -- the CMake project version still
said 0.1.0 while the repository was tagged v1.2.2.

build-native.yml drops the three iOS targets. They existed only for
tamga-swift's XCFramework pipeline, which went away when that repo replaced
its tamga-c binding with a native Swift implementation; confirmed no workflow
in the fleet still calls this one.

Three new examples cover the HTTP client, including the activation flow's
non-obvious part: creation does not check seat limits, so an integration that
stops after the create believes it succeeded and hands the user a copy they
are not entitled to.

Fuzzing: all six targets ran 120 seconds each under ASan and UBSan with no
findings. Coverage is 83.1% of lines.

+++ src/crypto/ed25519.c
@@ -30,11 +30,11 @@ static bool tamga_sc_is_canonical(const unsigned char s[32]) {
 }
 
 static void tamga_sc_load_l(uint32_t out[8]) {
-    int i;
-    for (i = 0; i < 8; i++) {
-        out[i] = (uint32_t)TAMGA_SC_L[i * 4] | ((uint32_t)TAMGA_SC_L[(i * 4) + 1] << 8) |
-                 ((uint32_t)TAMGA_SC_L[(i * 4) + 2] << 16) |
-                 ((uint32_t)TAMGA_SC_L[(i * 4) + 3] << 24);
+    size_t i;
+    for (i = 0u; i < 8u; i++) {
+        out[i] = (uint32_t)TAMGA_SC_L[i * 4u] | ((uint32_t)TAMGA_SC_L[(i * 4u) + 1u] << 8) |
+                 ((uint32_t)TAMGA_SC_L[(i * 4u) + 2u] << 16) |
+                 ((uint32_t)TAMGA_SC_L[(i * 4u) + 3u] << 24);
     }
 }
 
@@ -54,7 +54,7 @@ static void tamga_sc_reduce512(const unsigned char in[64], unsigned char out[32]
     uint32_t remainder[8];
     uint32_t l[8];
     int bit;
-    int i;
+    size_t i;
 
     memset(remainder, 0, sizeof(remainder));
     tamga_sc_load_l(l);
@@ -66,13 +66,13 @@ static void tamga_sc_reduce512(const unsigned char in[64], unsigned char out[32]
 
         /* remainder = remainder * 2 + next bit. The remainder stays below L,
          * so doubling it cannot exceed 2L < 2^254 and never overflows. */
-        for (i = 0; i < 8; i++) {
+        for (i = 0u; i < 8u; i++) {
             uint32_t next_carry = remainder[i] >> 31;
             remainder[i] = (remainder[i] << 1) | carry;
             carry = next_carry;
         }
 
-        for (i = 0; i < 8; i++) {
+        for (i = 0u; i < 8u; i++) {
             uint64_t diff = (uint64_t)remainder[i] - (uint64_t)l[i] - borrow;
             candidate[i] = (uint32_t)diff;
             borrow = (diff >> 63) & 1u;
@@ -82,11 +82,11 @@ static void tamga_sc_reduce512(const unsigned char in[64], unsigned char out[32]
         }
     }
 
-    for (i = 0; i < 8; i++) {
-        out[i * 4] = (unsigned char)remainder[i];
-        out[(i * 4) + 1] = (unsigned char)(remainder[i] >> 8);
-        out[(i * 4) + 2] = (unsigned char)(remainder[i] >> 16);
-        out[(i * 4) + 3] = (unsigned char)(remainder[i] >> 24);
+    for (i = 0u; i < 8u; i++) {
+        out[i * 4u] = (unsigned char)remainder[i];
+        out[(i * 4u) + 1u] = (unsigned char)(remainder[i] >> 8);
+        out[(i * 4u) + 2u] = (unsigned char)(remainder[i] >> 16);
+        out[(i * 4u) + 3u] = (unsigned char)(remainder[i] >> 24);
     }
 }
